+++ CHANGELOG.md
@@ -12,7 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests for "Genome sequencing".
 - Tests for "Roller coaster".
 - Tests for "Super computer".
-- Tests for "Surface".
+- Tests for "TAN network".
 
 ## [1.3.0] - 2022-03-02
 ### Added